+++ examples/minwebgl/curve_surface_rendering/src/main.rs
@@ -208,7 +208,7 @@ async fn setup_scene( gl : &WebGl2RenderingContext ) -> Result< GLTF, gl::WebglE
 {
   let window = web_sys::window().unwrap();
   let document =  window.document().unwrap();
-  let mut gltf = renderer::webgl::loaders::gltf::load( &document, "gltf/sphere.glb", &gl ).await?;
+  let mut gltf = renderer::webgl::loaders::gltf::load( &document, "static/gltf/sphere.glb", &gl ).await?;
 
   let earth = gltf.scenes[ 0 ].borrow().children.get( 1 ).unwrap().clone();
   let texture = create_texture( &gl, "textures/earth2.jpg" ).await;
@@ -327,7 +327,7 @@ async fn run() -> Result< (), gl::WebglError >
   camera.get_controls().borrow_mut().center = [ 0.0, 1.0, 0.0 ].into();
 
   let mut renderer = Renderer::new( &gl, canvas.width(), canvas.height(), 4 )?;
-  renderer.set_ibl( renderer::webgl::loaders::ibl::load( &gl, "environment_maps/gltf_viewer_ibl_unreal/", None ).await );
+  renderer.set_ibl( renderer::webgl::loaders::ibl::load( &gl, "static/environment_maps/gltf_viewer_ibl_unreal/", None ).await );
   let skybox = create_texture( &gl, "environment_maps/equirectangular_maps/space3.png" ).await
     .expect( "Failed to load skybox texture" );
   renderer.set_skybox( skybox.texture.borrow().source.clone() );